requirements.txt
librosa>=0.10.2,<1.0.0
noisereduce>=3.0.3,<4.0.0
numpy>=1.26.0,<3.0.0
openai>=1.66.0,<2.0.0
pandas>=2.2.0,<3.0.0
pydantic>=2.10.0,<3.0.0
pyannote.audio>=3.3.2,<4.0.0
python-dotenv>=1.0.1,<2.0.0
scikit-learn>=1.4.0,<2.0.0
soundfile>=0.12.1,<1.0.0
streamlit>=1.42.0,<2.0.0
tensorflow>=2.16.0,<3.0.0
tensorflow-hub>=0.16.1,<1.0.0
torch>=2.2.0,<3.0.0
webrtcvad-wheels>=2.0.14,<3.0.0

